+++ NEWS.md
@@ -1,6 +1,7 @@
 DALEX (development)
 ---------------------------------------------------------------
 * changed URLs in the DESCRIPTION as requested in ([#484](https://github.com/ModelOriented/DALEX/issues/484))
+* Fix model_info documentation ([#498](https://github.com/ModelOriented/DALEX/issues/498))
 
 DALEX 2.4.0
 ---------------------------------------------------------------